ignore doc/gh-pages directory, improve makedoc.bash
I set it up locally to be the working directory of a separate local repository
to set it up on another pc, run doc/makedoc.bash --gh-pages-setup

+++ doc/makedoc.bash
@@ -3,25 +3,53 @@
 cd "${0%/*}/"
 
 penview=$PWD/..
-
 ghpages=$PWD/gh-pages
 
-rm -rf "$ghpages/epydoc"
+_makedoc() {
+	if ! [ -d "$ghpages" ]
+	then
+		echo 'run "$0 --gh-pages-setup" first' >&2
+		exit 1
+	fi
+	rm -rf "$ghpages/epydoc"
+	mkdir -p "$ghpages/epydoc"
+	export PYTHONPATH="$penview/lib"
+	epydoc -v -o "$ghpages/epydoc" --no-private --docformat restructuredtext --name PenView --url http://p2000.github.com/penview/ "$penview"/*.py
+	read -p "git add commit push ? "
+	[[ "$REPLY" == [yY]* ]] || exit
+	cd "$ghpages"
+	git add .
+	git commit --amend -C HEAD
+	git push
+}
 
-mkdir -p "$ghpages/epydoc"
+_gh-pages-setup() {
+	mkdir "$ghpages"
+	cd "$ghpages"
+	git init
+	cat > .git/config <<-EOF
+		[core]
+		 	repositoryformatversion = 0
+		 	logallrefupdates = true
+		 	autocrlf = false
+		 	filemode = true
+		[remote "origin"]
+		 	url = ssh://git@github.com/P2000/penview.git
+		 	push = refs/heads/gh-pages:refs/heads/gh-pages
+		 	fetch = refs/heads/gh-pages:refs/remotes/origin/gh-pages
+		[branch "gh-pages"]
+		 	remote = origin
+		 	merge = refs/heads/gh-pages
+	EOF
+	git pull
+}
 
-export PYTHONPATH="$penview/lib"
-
-epydoc -v -o "$ghpages" --no-private --docformat restructuredtext --name PenView --url http://p2000.github.com/penview/ "$penview"/*.py
-
-read -p "git add commit push ? "
-
-[[ "$REPLY" == [yY]* ]] || exit
-
-cd "$pvpages"
-
-git add .
-
-git commit --amend -C HEAD
-
-git push
+case "$1"
+in
+	"") _makedoc
+	;;
+	--gh-pages-setup) _gh-pages-setup
+	;;
+	*) echo "usage: $0 [--gh-pages-setup]"
+	;;
+esac
